User menu // Data downloaders
The user menu is the backbone of the program where the user selects whether its a single team's matchup, all of the matchups, or see a player's predicted statline.

The downloaders are self explanatory, downloading all of the teams' stats for the year to use in the prediction functions. Also will download a player stats as needed as there are too many players to download at once.